pdf-editor/requirements.txt
PyPDF2>=3.0.0
PyMuPDF>=1.18.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0 
pytesseract